move some logic from autodiscovery to main eingine
